Modified: structure /lib/ with client/ and server/ directories and helpers.ts files in every component and page
