+++ app/entities/helpers.ts
@@ -16,23 +16,6 @@ export async function getEntities() {
 }
 
 /**
- * Displays subscription data from the backend
- * @returns Subscription data
- */
-export async function getSubscriptions() {
-  const url = `http://${location.hostname}:3004/api/getsubscriptions`;
-
-  try {
-    const response = await fetch(url);
-
-    if (!response.ok) throw new Error("Problems with the context broker");
-    return response;
-  } catch (error) {
-    if (error instanceof Error) throw new Error("Server problems");
-  }
-}
-
-/**
  * Create a subscription data and sends response to the backend
  * @param entity Entity data to set
  * @param valueNotificationWebhookUrl Notification webhook URL to set
@@ -120,32 +103,6 @@ export async function deleteEntityWithSubscriptions(id: string) {
       throw new Error(
         `Entity ${id} and their subscriptions were not deleted successfully`
       );
-
-    return response;
-  } catch (error) {
-    if (error instanceof Error) throw new Error(`Error: ${error.message}`);
-  }
-}
-
-/**
- * Delete subscription data and sends response to the backend
- * @param id Subscription ID to set
- * @returns A response of subscription data body for the backend
- */
-export async function deleteSubscription(id: string) {
-  const url = `http://${location.hostname}:3004/api/deletesubscription`;
-
-  try {
-    const response = await fetch(url, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ id }),
-    });
-
-    if (!response.ok)
-      throw new Error(`Entity ${id} was not deleted successfully`);
 
     return response;
   } catch (error) {
